requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
httpx
pypdf
groq
supabase
python-multipart
alembic
passlib[bcrypt]
python-jose[cryptography]
